backend/requirements.txt
fastapi==0.115.6
gunicorn==23.0.0
pydantic-settings==2.7.0
SQLAlchemy==2.0.36
uvicorn[standard]==0.32.1
